# Problem (Fix #1436): Performance of merkle library is not clear