### Benchmark result on laptop

#### Internal node fill rate

10000 stakings per version, commit 1000 versions in memory.

Internal node fill rate: 0.22 - 0.25, when the number of stakings is under 10M.

```
$ cargo run --example jellyfish_nodes -p chain-storage 1000 10000
...
internal/leaf: 65557 / 170000 = 0.3856294117647059
internal avg fill rate: 0.22457174672422472
...
internal/leaf: 807343 / 2060000 = 0.39191407766990294
internal avg fill rate: 0.22197365308177566
...
internal/leaf: 1090517 / 2840000 = 0.3839848591549296
internal avg fill rate: 0.2252667771341483
...
internal/leaf: 3333197 / 10000000 = 0.3333197
internal avg fill rate: 0.2500076503128978
```

#### In memory operations benchmark

```
$ cargo bench -p chain-storage
jellyfish, insert 256   time:   [4.6825 ms 4.6957 ms 4.7107 ms]
jellyfish, insert       time:   [30.993 us 31.144 us 31.313 us]
jellyfish, get_with_proof
                        time:   [15.605 us 15.721 us 15.877 us]
```

#### Disk usage

10000 stakings per version, no pruning.

```
$ cargo run --example disk_storage -p chain-storage /tmp/db 10 10000
$ du -sh /tmp/db
 37M	/tmp/db
$ rm -r /tmp/db
$ cargo run --example disk_storage -p chain-storage /tmp/db 100 10000
$ du -sh /tmp/db
 625M	/tmp/db
```

#### Disk get benchmark

Get staking from db with existing address:

```
$ DBPATH=/tmp/db STAKING_ADDRESS=0000000000000000000000000000000000000000  cargo run --example bench_jellyfish_disk -p chain-storage -- --bench
jellyfish_disk, get_with_proof
                        time:   [505.86 us 511.42 us 517.79 us]
```

Not exist address:
```
$ DBPATH=/tmp/db STAKING_ADDRESS=ffffffffffffffffffffffffffffffffffffffff cargo run --example bench_jellyfish_disk -p chain-storage -- --bench
jellyfish_disk, get_with_proof
                        time:   [370.42 us 375.35 us 383.60 us]
```