@yihuang 2020-04-21
> looks ok; for the disk get benchmark, it'll be more interesting to try with different addresses (not only zero-address)

Done