@leejw51crypto 2020-04-22
speed is not so fast , can i get speed with other ?