@leejw51crypto 2020-04-22
additional benchmark with starling is necessary