@leejw51crypto 2020-04-22
i think 
1 insert roughly 0.085 us