@tomtau 2020-04-22
> > looks ok; for the disk get benchmark, it'll be more interesting to try with different addresses (not only zero-address)
> 
> Done

I rather meant overall comparisons that it'll be looking up several different addresses in that microbenchmark -- to double check that the time for one particular existing address isn't some sort of the best edge case.

It's ok to merge now -- some further tests will be required later.

> speed is not so fast
> 
> also, In benchmark
> add os, cpu, ram
> 
> i also need benchmark with other merkletrie to compare

@leejw51crypto it seems ok for the account / staking storage, as there updates aren't that frequent  and for querying, it seems fine and one can load-balance. 

For UTXO commitments, it'll need more benchmarking and examination to find out whether the same trie could be used for that purpose, or something different would be preferable for lower overhead.