@tomtau 2020-04-22
(for UTXO commitment, one should experiment with a similar UTXO set size to Bitcoin)