@yihuang 2020-04-22
> (for UTXO commitment, one should experiment with a similar UTXO set size to Bitcoin)

Bitcoin UTxO size: 70M.
Jellyfish don't support deletion, I double they will support this in the future, so we might need to do this ourselves if we want to support UTxO.