@yihuang 2020-04-22
> > > looks ok; for the disk get benchmark, it'll be more interesting to try with different addresses (not only zero-address)
> > 
> > 
> > Done
> 
> I rather meant overall comparisons that it'll be looking up several different addresses in that microbenchmark -- to double check that the time for one particular existing address isn't some sort of the best edge case.
> 
> It's ok to merge now -- some further tests will be required later.

I checked with different addresses manually, the result doesn't change much, because the key is a further hashing of the address, so it will get randomized.